Move v4 tests
