+++ plugins/node/opentelemetry-instrumentation-express/test/v4/fixtures/use-express.mjs
@@ -15,16 +15,16 @@
  */
 
 // Use express from an ES module:
-//    node --experimental-loader=@opentelemetry/instrumentation/hook.mjs use-express-regex.mjs
+//    node --experimental-loader=@opentelemetry/instrumentation/hook.mjs use-express.mjs
 
 import { promisify } from 'util';
 import { createTestNodeSdk } from '@opentelemetry/contrib-test-utils';
 
 import { HttpInstrumentation } from '@opentelemetry/instrumentation-http';
-import { ExpressInstrumentation } from '../../build/src/index.js';
+import { ExpressInstrumentation } from '../../../build/src/index.js';
 
 const sdk = createTestNodeSdk({
-  serviceName: 'use-express-regex',
+  serviceName: 'use-express',
   instrumentations: [new HttpInstrumentation(), new ExpressInstrumentation()],
 });
 sdk.start();
@@ -43,26 +43,8 @@ app.use(async function simpleMiddleware(req, res, next) {
   next();
 });
 
-app.get(
-  [
-    '/test/arr/:id',
-    /\/test\/arr[0-9]*\/required(path)?(\/optionalPath)?\/(lastParam)?/,
-  ],
-  (_req, res) => {
-    res.send({ response: 'response' });
-  }
-);
-
-app.get(/\/test\/regex/, (_req, res) => {
-  res.send({ response: 'response 2' });
-});
-
-app.get(['/test/array1', /\/test\/array[2-9]/], (_req, res) => {
-  res.send({ response: 'response 3' });
-});
-
-app.get(['/test', '6', /test/], (_req, res) => {
-  res.send({ response: 'response 4' });
+app.get('/post/:id', (req, res) => {
+  res.send(`Post id: ${req.params.id}`);
 });
 
 const server = http.createServer(app);
@@ -70,15 +52,12 @@ await new Promise(resolve => server.listen(0, resolve));
 const port = server.address().port;
 
 await new Promise(resolve => {
-  http.get(
-    `http://localhost:${port}${process.env.TEST_REGEX_ROUTE}`,
-    res => {
-      res.resume();
-      res.on('end', () => {
-        resolve();
-      });
-    }
-  );
+  http.get(`http://localhost:${port}/post/0`, (res) => {
+    res.resume();
+    res.on('end', () => {
+      resolve();
+    });
+  })
 });
 
 await new Promise(resolve => server.close(resolve));
